before modify similarity of concept to rerank/hybrid
